vehicle speed changes

+++ PavlokDrivingApp/app/src/main/java/edu/bu/cs591/ateam/pavlokdrivingapp/MainActivity.java
@@ -297,7 +297,7 @@ public class MainActivity extends AppCompatActivity {
 
         @Override
         public void onLocationChanged(Location location) {
-            SpeedCheckTask.vehicleSpeed = (int) location.getSpeed();
+            SpeedCheckTask.vehicleSpeed = location.getSpeed();
         }
 
         @Override
